refactor: Move Scrollbar into separate file
